requirements.txt
pillow>=10.0.0
numpy>=1.24.0
zstandard>=0.22.0
customtkinter>=5.2.0
